added read write erase and all base driver functionality

+++ Components/Flash/Core/flash_low_api.c
@@ -44,7 +44,7 @@ flash_status_t flash_cmd_adr(flash_handle* flash, uint8_t cmd, uint32_t adr)
 	return retcode;
 }
 
-flash_status_t flash_transmit_data(flash_handle* flash, uint8_t cmd, uint32_t adr , const uint8_t* const data , uint16_t len)
+flash_status_t flash_transmit_data(flash_handle* flash, uint8_t cmd, uint32_t adr , const uint8_t* const data , uint32_t len)
 {
 	flash_status_t retcode = FLASH_OK;
 	if(NULL == flash || NULL == data)
@@ -74,7 +74,7 @@ flash_status_t flash_transmit_data(flash_handle* flash, uint8_t cmd, uint32_t ad
 	return retcode;
 }
 
-flash_status_t flash_receive_data(flash_handle* flash, uint8_t cmd, uint32_t adr ,uint8_t* data , uint16_t len)
+flash_status_t flash_receive_data(flash_handle* flash, uint8_t cmd, uint32_t adr ,uint8_t* data , uint32_t len)
 {
 	flash_status_t retcode = FLASH_OK;
 	if(NULL == flash || NULL == data)
