+++ Components/Flash/Inc/flash_low_api.h
@@ -14,10 +14,10 @@ flash_status_t flash_transmit_data(flash_handle* flash,
 								  uint8_t cmd,
 								  uint32_t adr ,
 								  const uint8_t* const data ,
-								  uint16_t len);
+								  uint32_t len);
 
 flash_status_t flash_receive_data(flash_handle* flash,
 								  uint8_t cmd,
 								  uint32_t adr ,
 								  uint8_t* data ,
-								  uint16_t len);
+								  uint32_t len);
